chore: finalize rename to file2ai and reorganize repository
