+++ CONTRIBUTING.md
@@ -1,13 +1,16 @@
-# Contributing to git2txt
+# Contributing to file2ai
 
-Thank you for your interest in contributing to git2txt! This document provides guidelines and instructions for contributing to the project.
+Thank you for your interest in contributing to file2ai! This document provides guidelines and instructions for contributing to the project.
+
+### Deprecation Notice
+The `git2txt` command is deprecated and will be removed in a future version. Please use `file2ai` instead. A compatibility layer is currently maintained for backward compatibility, but will be removed in version 2.0.0.
 
 ## Development Setup
 
 1. Clone the repository:
 ```bash
-git clone https://github.com/michaeljabbour/git2txt.git
-cd git2txt
+git clone https://github.com/michaeljabbour/file2ai.git
+cd file2ai
 ```
 
 2. Create and activate a virtual environment:
