+++ README.md
@@ -1,6 +1,6 @@
-# git2txt Exporter
+# file2ai
 
-This is a simple Python script that extracts all of the text in files from either a local directory or remote GitHub repository (by cloning) or a local directory. The script extracts the context of the files recursively and flattens them into a single file that is easy for ingestion into a LLM. 
+A comprehensive tool for file conversion, text extraction, and AI processing. This tool can extract text from Git repositories, local directories, and convert files between various formats, making it easy to prepare data for AI model ingestion.
 
 ## How It Works
 
@@ -48,12 +48,12 @@ my_project/
 
 **Command**:
 ```bash
-python git2txt.py --local-dir my_project
+python file2ai.py --local-dir my_project
 ```
 
 **Output Snippet**:
 ```
-Generated by git2txt
+Generated by file2ai
 ================================================================================
 
 Directory Structure:
@@ -86,18 +86,18 @@ Last Commit: ...
 - Automatically extracts base repository URL from deep URLs
 - Supports subdirectory processing with --repo-url-sub flag (must come before URL)
 - All flags must be specified before the repository URL
-- Run `python git2txt.py --help` to see all available options
+- Run `python file2ai.py --help` to see all available options
 
 ## Installation
 
-There are two ways to install and use git2txt:
+There are two ways to install and use file2ai:
 
 ### Option A: Development Installation
 
 1. Clone the repository:
 ```bash
-git clone https://github.com/michaeljabbour/git2txt.git
-cd git2txt
+git clone https://github.com/michaeljabbour/file2ai.git
+cd file2ai
 ```
 
 2. (Optional) Create and activate a virtual environment:
@@ -110,76 +110,114 @@ source venv/bin/activate  # or .\venv\Scripts\activate on Windows
 ```bash
 pip install -e .
 ```
-This installs git2txt as a package and includes development dependencies.
+This installs file2ai as a package and includes development dependencies.
 
 ### Option B: Direct Script Usage
 
 Simply clone and run:
 ```bash
-git clone https://github.com/michaeljabbour/git2txt.git
-cd git2txt
-python git2txt.py --repo-url https://github.com/owner/repo.git
+git clone https://github.com/michaeljabbour/file2ai.git
+cd file2ai
+python file2ai.py --repo-url https://github.com/owner/repo.git
 ```
 
 Note: The script will automatically install GitPython if it's not already available.
 
 ## Quick Start Usage
 
-You can run `git2txt.py` directly without installing anything else, as the script handles dependencies automatically. Below are common usage patterns:
+You can run `file2ai.py` directly without installing anything else, as the script handles dependencies automatically. Below are common usage patterns:
 
 ### 1. Export From a Remote GitHub Repo
 
 Basic usage:
 ```bash
-python git2txt.py --repo-url https://github.com/owner/repo.git
+python file2ai.py --repo-url https://github.com/owner/repo.git
 ```
 
 Deep URL handling:
 ```bash
 # Export from repository root (automatically extracts base repository):
-python git2txt.py https://github.com/owner/repo/pulls
-python git2txt.py https://github.com/owner/repo/tree/main/docs
-python git2txt.py https://github.com/owner/repo/issues
+python file2ai.py https://github.com/owner/repo/pulls
+python file2ai.py https://github.com/owner/repo/tree/main/docs
+python file2ai.py https://github.com/owner/repo/issues
 
 # Export from specific subdirectory (use --repo-url-sub flag):
-python git2txt.py --repo-url-sub https://github.com/owner/repo/tree/main/docs
+python file2ai.py --repo-url-sub https://github.com/owner/repo/tree/main/docs
 ```
 
 Optional: Specify a branch or commit (flags must come before URL):
 ```bash
-python git2txt.py --branch dev https://github.com/owner/repo.git
+python file2ai.py --branch dev https://github.com/owner/repo.git
 ```
 
 Private Repos: Provide your Personal Access Token (flags must come before URL):
 ```bash
-python git2txt.py --token YOUR_TOKEN https://github.com/owner/private-repo.git
+python file2ai.py --token YOUR_TOKEN https://github.com/owner/private-repo.git
 ```
 
 Skip Removing the Clone: Keep the cloned repo after export (for debugging):
 ```bash
-python git2txt.py --skip-remove https://github.com/owner/repo.git
+python file2ai.py --skip-remove https://github.com/owner/repo.git
 ```
 
 ### 2. Export From a Local Directory
 
 ```bash
-python git2txt.py --local-dir /path/to/local/project
+python file2ai.py --local-dir /path/to/local/project
 ```
 
-- If the directory contains a `.git` folder, git2txt will attempt to gather commit info for each file
+- If the directory contains a `.git` folder, file2ai will attempt to gather commit info for each file
 - If not, it still processes files but omits commit data
 
 ### 3. Custom Output Filename
 
 ```bash
-python git2txt.py --output-file my_export.txt https://github.com/owner/repo.git
+python file2ai.py --output-file my_export.txt https://github.com/owner/repo.git
 ```
 
 This places the file in `exports/my_export.txt`.
 
+### 4. Multi-Format Document Conversion
+
+The `convert` subcommand enables conversion between various document formats:
+
+```bash
+# Convert Word document to PDF
+python file2ai.py convert --input doc.docx --format pdf
+
+# Convert PowerPoint to images (one per slide)
+python file2ai.py convert --input pres.pptx --format image
+
+# Convert Excel to text (CSV)
+python file2ai.py convert --input sheet.xlsx --format text
+
+# Convert PDF to text with custom options
+python file2ai.py convert --input doc.pdf --format text --brightness 1.2 --contrast 1.1
+```
+
+Supported input formats:
+- Microsoft Word (DOC, DOCX)
+- Microsoft PowerPoint (PPT, PPTX)
+- Microsoft Excel (XLS, XLSX)
+- PDF documents
+- HTML/MHTML web pages
+
+Output format options:
+- Text (TXT, CSV)
+- PDF
+- Images (PNG, JPEG)
+- MHTML
+
+Additional options:
+- `--brightness`: Adjust image brightness (0.5-2.0)
+- `--contrast`: Adjust image contrast (0.5-2.0)
+- `--resolution`: Set output resolution for images (DPI)
+- `--pages`: Specify page range (e.g., "1-5,7,9-12")
+- `--pdf-compliance`: Set PDF compliance level (e.g., PDF/A)
+
 ## Development
 
-After cloning, you can edit `git2txt.py` or the supporting modules. We recommend using a virtual environment and the pytest library for testing.
+After cloning, you can edit `file2ai.py` or the supporting modules. We recommend using a virtual environment and the pytest library for testing.
 
 ### Code Style
