+++ test_script.sh
@@ -2,7 +2,7 @@
 
 # test_script.sh
 # --------------------------------------------------
-# Automated validation script for git2txt
+# Automated validation script for file2ai
 #
 # Usage:
 #   ./test_script.sh
@@ -13,7 +13,7 @@
 #
 # This script:
 #   1) Creates a clean virtual environment
-#   2) Installs git2txt and test dependencies
+#   2) Installs file2ai and test dependencies
 #   3) Runs pytest with coverage reporting
 #   4) Performs export tests:
 #      - Local directory export
@@ -30,7 +30,7 @@
 #   1 - Test or validation failure
 #
 # Example Output Files:
-#   exports/git2txt_export.txt    - Local directory export
+#   exports/file2ai_export.txt    - Local directory export
 #   exports/docling_export.txt    - Remote repository export
 #   exports/docling_export(1).txt - Subdirectory export
 #
@@ -76,7 +76,7 @@ source venv/bin/activate
 log_success "Virtual environment created and activated"
 
 # 3) Install package and dependencies
-log_info "Installing git2txt in editable mode..."
+log_info "Installing file2ai in editable mode..."
 pip install --upgrade pip
 pip install -e .
 pip install pytest pytest-cov
@@ -84,31 +84,31 @@ log_success "Installation complete"
 
 # 4) Run tests with coverage
 log_info "Running tests with coverage..."
-pytest --cov=git2txt || log_error "Tests failed!"
+pytest --cov=file2ai || log_error "Tests failed!"
 log_success "Tests passed"
 
 # 5) Test local directory export
 log_info "Testing local directory export..."
-python git2txt.py --local-dir . || log_error "Local export failed!"
+python file2ai.py --local-dir . || log_error "Local export failed!"
 
 # 6) Test normal remote repo export
 log_info "Testing normal remote repo export..."
-python git2txt.py --repo-url https://github.com/michaeljabbour/git2txt || log_error "Remote repo export failed!"
+python file2ai.py --repo-url https://github.com/michaeljabbour/file2ai || log_error "Remote repo export failed!"
 
 # 7) Test subdir/extra path export
 log_info "Testing subdir/extra path export..."
-python git2txt.py --repo-url-sub https://github.com/michaeljabbour/git2txt/pulls || log_error "Subdir export failed!"
+python file2ai.py --repo-url-sub https://github.com/michaeljabbour/file2ai/pulls || log_error "Subdir export failed!"
 
 # 8) Validate outputs
 log_info "Validating output files..."
 
 # Check local export
-txt_file="exports/git2txt_export.txt"
+txt_file="exports/file2ai_export.txt"
 if [ -f "$txt_file" ]; then
     log_info "Local directory export found"
     
     # Basic content validation
-    if grep -q "Generated by git2txt" "$txt_file" && \
+    if grep -q "Generated by file2ai" "$txt_file" && \
        grep -q "Directory Structure:" "$txt_file" && \
        grep -q "=" "$txt_file"; then
         log_success "Local export structure looks correct"
@@ -128,12 +128,12 @@ else
 fi
 
 # Check remote repo export
-txt_file="exports/git2txt_export.txt"
+txt_file="exports/file2ai_export.txt"
 if [ -f "$txt_file" ]; then
     log_info "Remote repo text export found"
     
     # Basic content validation
-    if grep -q "Generated by git2txt" "$txt_file" && \
+    if grep -q "Generated by file2ai" "$txt_file" && \
        grep -q "Directory Structure:" "$txt_file" && \
        grep -q "=" "$txt_file"; then
         log_success "Text export structure looks correct"
@@ -153,10 +153,10 @@ else
 fi
 
 # Check subdir export
-subdir_file="exports/git2txt_export(1).txt"
+subdir_file="exports/file2ai_export(1).txt"
 if [ -f "$subdir_file" ]; then
     log_info "Subdir export found"
-    if grep -q "Generated by git2txt" "$subdir_file"; then
+    if grep -q "Generated by file2ai" "$subdir_file"; then
         log_success "Subdir export structure looks correct"
     else
         log_error "Subdir export structure is invalid"
